chore(download): deleted old support and moved packages
