+++ infrastructure/http/src/main/java/top/chiloven/lukosbot2/util/DownloadUtils.kt
@@ -17,33 +17,32 @@
  */
 package top.chiloven.lukosbot2.util
 
+import top.chiloven.lukosbot2.util.download.*
 import java.io.IOException
 import java.net.URI
+import java.nio.file.Files
 import java.nio.file.Path
 
 /**
- * Compatibility bridge for the previous download utility API.
- *
- * New project code should call [DownloadClient] directly.
+ * Public facade for resilient HTTP downloads.
  */
-@Deprecated(
-    message = "This object was remained for compatibility. Please use DownloadClient instead.",
-    replaceWith = ReplaceWith(
-        expression = "DownloadClient",
-        imports = ["top.chiloven.lukosbot2.util.DownloadClient"]
-    )
-)
 object DownloadUtils {
 
-    const val DEFAULT_MAX_CONCURRENT_FILES: Int = DownloadClient.DEFAULT_MAX_CONCURRENT_FILES
-    const val DEFAULT_CHUNK_THREADS: Int = DownloadClient.DEFAULT_CHUNK_THREADS
-    const val DEFAULT_MIN_SIZE_FOR_CHUNKING_BYTES: Long = DownloadClient.DEFAULT_MIN_SIZE_FOR_CHUNKING_BYTES
-    const val DEFAULT_MIN_PART_SIZE_BYTES: Long = DownloadClient.DEFAULT_MIN_PART_SIZE_BYTES
-    const val DEFAULT_MAX_RETRIES: Int = DownloadClient.DEFAULT_MAX_RETRIES
-    const val DEFAULT_RETRY_BASE_DELAY_MS: Long = DownloadClient.DEFAULT_RETRY_BASE_DELAY_MS
-    const val DEFAULT_RETRY_MAX_DELAY_MS: Long = DownloadClient.DEFAULT_RETRY_MAX_DELAY_MS
-    const val DEFAULT_RETRY_AFTER_CAP_MS: Long = DownloadClient.DEFAULT_RETRY_AFTER_CAP_MS
-    const val DEFAULT_PROGRESS_LOG_INTERVAL_MS: Long = DownloadClient.DEFAULT_PROGRESS_LOG_INTERVAL_MS
+    const val DEFAULT_MAX_CONCURRENT_FILES: Int = DownloadDefaults.DEFAULT_MAX_CONCURRENT_FILES
+    const val DEFAULT_CHUNK_THREADS: Int = DownloadDefaults.DEFAULT_CHUNK_THREADS
+    const val DEFAULT_MIN_SIZE_FOR_CHUNKING_BYTES: Long = DownloadDefaults.DEFAULT_MIN_SIZE_FOR_CHUNKING_BYTES
+    const val DEFAULT_MIN_PART_SIZE_BYTES: Long = DownloadDefaults.DEFAULT_MIN_PART_SIZE_BYTES
+    const val DEFAULT_MAX_RETRIES: Int = DownloadDefaults.DEFAULT_MAX_RETRIES
+    const val DEFAULT_RETRY_BASE_DELAY_MS: Long = DownloadDefaults.DEFAULT_RETRY_BASE_DELAY_MS
+    const val DEFAULT_RETRY_MAX_DELAY_MS: Long = DownloadDefaults.DEFAULT_RETRY_MAX_DELAY_MS
+    const val DEFAULT_RETRY_AFTER_CAP_MS: Long = DownloadDefaults.DEFAULT_RETRY_AFTER_CAP_MS
+    const val DEFAULT_PROGRESS_LOG_INTERVAL_MS: Long = DownloadDefaults.DEFAULT_PROGRESS_LOG_INTERVAL_MS
+
+    private val http = DownloadHttp()
+    private val retryPolicyFactory: (Int) -> RetryPolicy = { RetryPolicy.default(it) }
+    private val singleFileDownloader = SingleFileDownloader(http, retryPolicyFactory)
+    private val rangeDownloader = RangeDownloader(http, singleFileDownloader, retryPolicyFactory)
+    private val batchDownloader = BatchDownloader(singleFileDownloader, rangeDownloader)
 
     @JvmStatic
     @Throws(IOException::class)
@@ -52,12 +51,14 @@ object DownloadUtils {
         dir: Path,
         headers: Map<String, String>?,
         timeoutMs: Int,
-    ): BatchResult = DownloadClient.downloadAllToDir(
-        items = items.toClientItems(),
+    ): BatchResult = batchDownloader.download(
+        items = items,
         dir = dir,
         headers = headers,
-        timeoutMs = timeoutMs
-    ).toLegacyResult()
+        timeoutMs = timeoutMs,
+        namingMode = BatchDownloader.BatchNamingMode.FLAT_FILES,
+        options = BatchDownloader.BatchDownloadOptions()
+    )
 
     @JvmStatic
     @JvmOverloads
@@ -70,15 +71,14 @@ object DownloadUtils {
         maxConcurrentFiles: Int = DEFAULT_MAX_CONCURRENT_FILES,
         chunkThreadsPerFile: Int = 1,
         maxRetries: Int = DEFAULT_MAX_RETRIES,
-    ): BatchResult = DownloadClient.downloadAllToDirConcurrent(
-        items = items.toClientItems(),
+    ): BatchResult = batchDownloader.download(
+        items = items,
         dir = dir,
         headers = headers,
         timeoutMs = timeoutMs,
-        maxConcurrentFiles = maxConcurrentFiles,
-        chunkThreadsPerFile = chunkThreadsPerFile,
-        maxRetries = maxRetries
-    ).toLegacyResult()
+        namingMode = BatchDownloader.BatchNamingMode.FLAT_FILES,
+        options = BatchDownloader.BatchDownloadOptions(maxConcurrentFiles, chunkThreadsPerFile, maxRetries)
+    )
 
     @JvmStatic
     @JvmOverloads
@@ -91,15 +91,14 @@ object DownloadUtils {
         maxConcurrentFiles: Int = DEFAULT_MAX_CONCURRENT_FILES,
         chunkThreadsPerFile: Int = DEFAULT_CHUNK_THREADS,
         maxRetries: Int = DEFAULT_MAX_RETRIES,
-    ): BatchResult = DownloadClient.downloadNamedUrlsToDirConcurrent(
-        items = items.toClientItems(),
+    ): BatchResult = batchDownloader.download(
+        items = items,
         dir = dir,
         headers = headers,
         timeoutMs = timeoutMs,
-        maxConcurrentFiles = maxConcurrentFiles,
-        chunkThreadsPerFile = chunkThreadsPerFile,
-        maxRetries = maxRetries
-    ).toLegacyResult()
+        namingMode = BatchDownloader.BatchNamingMode.RELATIVE_PATHS,
+        options = BatchDownloader.BatchDownloadOptions(maxConcurrentFiles, chunkThreadsPerFile, maxRetries)
+    )
 
     @JvmStatic
     @JvmOverloads
@@ -111,7 +110,7 @@ object DownloadUtils {
         timeoutMs: Int,
         maxRetries: Int = DEFAULT_MAX_RETRIES,
     ) {
-        DownloadClient.downloadToFile(url, targetFile, headers, timeoutMs, maxRetries)
+        singleFileDownloader.downloadToFile(url, targetFile, headers, timeoutMs, maxRetries)
     }
 
     @JvmStatic
@@ -124,7 +123,12 @@ object DownloadUtils {
         headers: Map<String, String>?,
         timeoutMs: Int,
         maxRetries: Int = DEFAULT_MAX_RETRIES,
-    ): Path = DownloadClient.downloadToDir(url, dir, fileName, headers, timeoutMs, maxRetries)
+    ): Path {
+        Files.createDirectories(dir)
+        val target = DownloadNaming.resolveFlatTarget(dir, fileName)
+        downloadToFile(url, target, headers, timeoutMs, maxRetries)
+        return target
+    }
 
     @JvmStatic
     @JvmOverloads
@@ -139,7 +143,7 @@ object DownloadUtils {
         minPartSizeBytes: Long = DEFAULT_MIN_PART_SIZE_BYTES,
         maxRetries: Int = DEFAULT_MAX_RETRIES,
     ) {
-        DownloadClient.downloadToFileFast(
+        rangeDownloader.downloadToFileFast(
             url = url,
             targetFile = targetFile,
             headers = headers,
@@ -162,17 +166,24 @@ object DownloadUtils {
         timeoutMs: Int,
         chunkThreads: Int = DEFAULT_CHUNK_THREADS,
         maxRetries: Int = DEFAULT_MAX_RETRIES,
-    ): Path = DownloadClient.downloadToDirFast(url, dir, fileName, headers, timeoutMs, chunkThreads, maxRetries)
+    ): Path {
+        Files.createDirectories(dir)
+        val target = DownloadNaming.resolveFlatTarget(dir, fileName)
+        downloadToFileFast(
+            url = url,
+            targetFile = target,
+            headers = headers,
+            timeoutMs = timeoutMs,
+            chunkThreads = chunkThreads,
+            minSizeForChunking = DEFAULT_MIN_SIZE_FOR_CHUNKING_BYTES,
+            minPartSizeBytes = DEFAULT_MIN_PART_SIZE_BYTES,
+            maxRetries = maxRetries
+        )
+        return target
+    }
 
     @JvmStatic
-    fun sanitizeFileName(name: String?): String = DownloadClient.sanitizeFileName(name)
-
-    private fun List<NamedUrl?>?.toClientItems(): List<DownloadClient.NamedUrl?>? =
-        this?.map { item -> item?.toClientItem() }
-
-    private fun NamedUrl.toClientItem(): DownloadClient.NamedUrl = DownloadClient.NamedUrl(name, url)
-
-    private fun DownloadClient.BatchResult.toLegacyResult(): BatchResult = BatchResult(ok, failed)
+    fun sanitizeFileName(name: String?): String = DownloadNaming.flatFileName(name)
 
     data class NamedUrl(
         val name: String,
