+++ infrastructure/http/src/main/java/top/chiloven/lukosbot2/util/download/DownloadHttp.kt
@@ -15,7 +15,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package top.chiloven.lukosbot2.util
+package top.chiloven.lukosbot2.util.download
 
 import okhttp3.Headers
 import okhttp3.OkHttpClient
@@ -23,6 +23,7 @@ import okhttp3.Request
 import okhttp3.Response
 import org.apache.logging.log4j.LogManager
 import top.chiloven.lukosbot2.Constants
+import top.chiloven.lukosbot2.util.OkHttpUtils
 import java.io.IOException
 import java.net.URI
 import java.util.concurrent.TimeUnit
