+++ infrastructure/http/src/main/java/top/chiloven/lukosbot2/util/download/HttpStatusException.kt
@@ -15,7 +15,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package top.chiloven.lukosbot2.util
+package top.chiloven.lukosbot2.util.download
 
 import java.io.IOException
